move to returnn/hdf.py + use soundfile
